Feature/e1 (#1)
* Implement basic kinematic chain

Also update project name to RSD_Exercise

* Split solution from student template

Also re-added the launch.json to be used for debugging (at least ex1)
since there is a bug wiht matplotlib, qt and debugpy. It is easily
solveable, but needs a "guiEventLoop" flag in the vscode settings.

* Added ReadMe to the project

* fixup! Split solution from student template

* fixup! Split solution from student template

* Lowercase project name

This is the pythonic way

* Improved code clarity

Also added the joint type to the plot

Added task description

* fixup! Added ReadMe to the project