Fix bug in solution when calculating the homogenous global transformation matrix (#2)
Fix bug in solution when calculating the homogenous global transformation matrix. In the old version, the homogenous transform was set to identity when no parent was found. This was neglecting the first joint in the chain. Now, instead of returning identity, the parent is assumed to be Identity and the local joint transform is still applied.
In conjunction, the default joints in the playground have been updated to better visualize the different joint axes (and look nicer).

This PR fixes the jumpy (and rather random) animation.

In addition, the axis parameter of Joint3D has been further clarified